Add helper to format graph
